Support aggregating Kover reports without a need for a dedicated module
