+++ example/kotlinlib/linting/4-kover/bar/src/BarKotlin.kt
@@ -1,0 +1,17 @@
+package bar
+
+fun action(one: Boolean, two: Boolean): String {
+    return if (one) {
+        if (two) {
+            "one, two"
+        } else {
+            "one"
+        }
+    } else {
+        if (two) {
+            "two"
+        } else {
+            "nothing"
+        }
+    }
+}
