+++ example/kotlinlib/linting/4-kover/bar/test/src/BarTests.kt
@@ -1,0 +1,10 @@
+package bar
+
+import io.kotest.core.spec.style.FunSpec
+import io.kotest.matchers.shouldBe
+
+class BarTests : FunSpec({
+    test("kotlin - success") {
+        action(true, true) shouldBe "one, two"
+    }
+})
