+++ example/kotlinlib/linting/4-kover/build.mill
@@ -5,11 +5,7 @@ import kotlinlib.contrib.kover.KoverModule
 
 object `package` extends RootModule with KotlinModule with KoverModule {
 
-  def kotlinVersion = "1.9.24"
-
-  override def koverVersion = "0.8.3"
-
-  object test extends KotlinModuleTests with TestModule.Junit5 with KoverTests {
+  trait KotestTests extends TestModule.Junit5 {
     override def forkArgs: T[Seq[String]] = Task {
       super.forkArgs() ++ Seq("-Dkotest.framework.classpath.scanning.autoscan.disable=true")
 
@@ -17,6 +13,17 @@ object `package` extends RootModule with KotlinModule with KoverModule {
     override def ivyDeps = super.ivyDeps() ++ Agg(
       ivy"io.kotest:kotest-runner-junit5-jvm:5.9.1"
     )
+  }
+
+  def kotlinVersion = "1.9.24"
+
+  override def koverVersion = "0.8.3"
+
+  object test extends KotlinModuleTests with KotestTests with KoverTests
+
+  object bar extends KotlinModule with KoverModule {
+    def kotlinVersion = "1.9.24"
+    object test extends KotlinModuleTests with KotestTests with KoverTests
   }
 }
 
@@ -32,6 +39,8 @@ object `package` extends RootModule with KotlinModule with KoverModule {
 // must first run the test.
 // `./mill test` then `./mill show kover.htmlReport` and get your
 // coverage in HTML format.
+// Also reports for all modules can be collected in a single place by
+// running `./mill show mill.kotlinlib.contrib.kover.Kover/htmlReportAll`.
 
 /** Usage
 
@@ -51,6 +60,14 @@ kover.xmlReport
 ...out/kover/htmlReport.dest/kover-report...
 
 > cat out/kover/htmlReport.dest/kover-report/index.html
+...
+...Kover HTML Report: Overall Coverage Summary...
+
+> ./mill show mill.kotlinlib.contrib.kover.Kover/htmlReportAll # collect reports from all modules
+...
+...out/mill/kotlinlib/contrib/kover/Kover/htmlReportAll.dest/kover-report...
+
+> cat out/mill/kotlinlib/contrib/kover/Kover/htmlReportAll.dest/kover-report/index.html
 ...
 ...Kover HTML Report: Overall Coverage Summary...
