+++ kotlinlib/test/src/mill/kotlinlib/contrib/kover/KoverModuleTests.scala
@@ -39,8 +39,6 @@ object KoverModuleTests extends TestSuite {
       def kotlinVersion = "1.9.24"
       object test extends KotlinModuleTests with module.KotestTestModule
     }
-
-    object kover extends KoverReportModule
   }
 
   def tests: Tests = Tests {
@@ -61,14 +59,13 @@ object KoverModuleTests extends TestSuite {
             )
         )
 
-      val Right(result) = eval(module.kover.xmlReportAll(eval.evaluator))
+      val Right(result) = eval(Kover.xmlReportAll(eval.evaluator))
 
       val xmlReportPath = result.value.path
       assert(os.exists(xmlReportPath))
-      val relPath = xmlReportPath.segments.toVector.dropRight(1).takeRight(3)
-      assert(relPath.head == "out")
-      assert(relPath(1) == "kover")
-      assert(relPath(2) == "xmlReportAll.dest")
+      val relPath = xmlReportPath.segments.toVector.takeRight(2)
+      assert(relPath.head == "xmlReportAll.dest")
+      assert(relPath.last == "kover-report.xml")
 
       val xmlReport = XML.loadFile(xmlReportPath.toString)
 
@@ -81,7 +78,7 @@ object KoverModuleTests extends TestSuite {
 
     }
 
-    test("report") {
+    test("report-xml") {
 
       val eval = UnitTester(module, resourcePath)
 
@@ -91,12 +88,14 @@ object KoverModuleTests extends TestSuite {
 
       val xmlReportPath = result.value.path
       assert(os.exists(xmlReportPath))
+      assert(os.isFile(xmlReportPath))
 
       // drop report name
-      val relPath = xmlReportPath.segments.toVector.dropRight(1).takeRight(3)
+      val relPath = xmlReportPath.segments.toVector.takeRight(4)
       assert(relPath.head == "foo")
       assert(relPath(1) == "kover")
       assert(relPath(2) == "xmlReport.dest")
+      assert(relPath(3) == "kover-report.xml")
 
       val xmlReport = XML.loadFile(xmlReportPath.toString)
 
@@ -107,6 +106,28 @@ object KoverModuleTests extends TestSuite {
       assert(packageNameChildNode(xmlReport, "bar").isEmpty)
       assert(packageNameChildNode(xmlReport, "qux").isEmpty)
 
+    }
+
+    test("report-html") {
+
+      val eval = UnitTester(module, resourcePath)
+
+      val Right(_) = eval(module.foo.test.test())
+
+      val Right(result) = eval(module.foo.kover.htmlReport())
+
+      val htmlReportPath = result.value.path
+      assert(os.exists(htmlReportPath))
+      assert(os.isDir(htmlReportPath))
+      assert(os.walk(htmlReportPath)
+        .exists(p => p.ext == "html"))
+
+      // drop report name
+      val relPath = htmlReportPath.segments.toVector.takeRight(4)
+      assert(relPath.head == "foo")
+      assert(relPath(1) == "kover")
+      assert(relPath(2) == "htmlReport.dest")
+      assert(relPath(3) == "kover-report")
     }
   }
